+++ changelog.md
@@ -7,6 +7,8 @@
 - Added full options menu
 - Added no respawn blink
 - Added level uncomplete
+- Added show hitboxes and show hitboxes on death
+- Added no static camera
 - Removed fps bypass since it was VERY broken, replaces with full options menu
 - Fixed Spider and Robot with icon effects
 - Added <cl>Gradient Pages</c> integration